+++ Profile/Shopware/DataSelection/WishlistDataSelection.php
@@ -1,0 +1,79 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SwagMigrationAssistant\Profile\Shopware\DataSelection;
+
+use SwagMigrationAssistant\Migration\DataSelection\DataSelectionInterface;
+use SwagMigrationAssistant\Migration\DataSelection\DataSelectionStruct;
+use SwagMigrationAssistant\Migration\MigrationContextInterface;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\CrossSellingDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\CustomerAttributeDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\CustomerDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\CustomerWishlistDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\MainVariantRelationDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ManufacturerAttributeDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\MediaFolderDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ProductAttributeDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ProductDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ProductOptionRelationDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ProductPriceAttributeDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\ProductPropertyRelationDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\PropertyGroupOptionDataSet;
+use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\TranslationDataSet;
+use SwagMigrationAssistant\Profile\Shopware\ShopwareProfileInterface;
+
+class WishlistDataSelection implements DataSelectionInterface
+{
+    public const IDENTIFIER = 'customerWishlists';
+
+    public function supports(MigrationContextInterface $migrationContext): bool
+    {
+        return $migrationContext->getProfile() instanceof ShopwareProfileInterface;
+    }
+
+    public function getData(): DataSelectionStruct
+    {
+        return new DataSelectionStruct(
+            self::IDENTIFIER,
+            $this->getDataSets(),
+            $this->getDataSetsRequiredForCount(),
+            'swag-migration.index.selectDataCard.dataSelection.customerWishlists',
+            220,
+            true
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDataSets(): array
+    {
+        return [
+            new MediaFolderDataSet(),
+            new ProductAttributeDataSet(),
+            new ProductPriceAttributeDataSet(),
+            new ManufacturerAttributeDataSet(),
+            new ProductDataSet(),
+            new PropertyGroupOptionDataSet(),
+            new ProductOptionRelationDataSet(),
+            new ProductPropertyRelationDataSet(),
+            new TranslationDataSet(),
+            new CrossSellingDataSet(),
+            new MainVariantRelationDataSet(),
+            new CustomerAttributeDataSet(),
+            new CustomerDataSet(),
+            new CustomerWishlistDataSet(),
+        ];
+    }
+
+    public function getDataSetsRequiredForCount(): array
+    {
+        return [
+            new CustomerWishlistDataSet(),
+        ];
+    }
+}